requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.3.4
billiard==3.3.0.23
bootstrap-admin==0.3.7.1
celery==3.1.24
Django==1.9.5
django-annoying==0.10.3
django-bower==5.2.0
django-crispy-forms==1.6.0
django-dashing==0.3.2
django-debug-toolbar==1.5
django-excel==0.0.6
django-filter==1.0.1
django-phonenumber-field==1.1.0
django-recaptcha==1.2.1
django-scheduler==0.8.2
django-tenant-schemas==1.5.8
djangorestframework==3.5.3
et-xmlfile==1.0.1
flower==0.9.1
icalendar==3.11.2
jdcal==1.3
kombu==3.0.37
Markdown==2.6.7
nltk==3.2.2
num2words==0.5.4
openpyxl==2.4.1
phonenumbers==7.7.2
phonenumberslite==7.7.2
Pillow==3.4.1
psycopg2==2.6.1
pyexcel==0.4.2
pyexcel-io==0.3.0
pyexcel-webio==0.0.10
pyexcel-xls==0.3.1
pyexcel-xlsx==0.3.0
python-dateutil==2.6.0
pytz==2016.7
reportlab==3.3.0
six==1.10.0
sqlparse==0.2.2
texttable==0.8.7
tornado==4.2
xlrd==1.0.0
XlsxWriter==0.9.3
xlwt==1.2.0
